refactor: 오브제 리팩토링
- Router Outlet 적용
  - 오브제 상세조회
  - 오브제 채팅
  - 오브제 통화
